Consolidate extension invocation ownership
