+++ crates/homeboy-core/src/extension/invoke/action.rs
@@ -5,9 +5,9 @@ use homeboy_core::project;
 use homeboy_core::server::http::ApiClient;
 use homeboy_engine_primitives::validation;
 
+use super::scope::ExtensionScope;
 use super::{build_action_env, execute_extension_command, ExtensionExecutionMode};
 use crate::extension::catalog::load_extension;
-use crate::extension_scope::ExtensionScope;
 use homeboy_extension_contract::action_types::{ActionType, HttpMethod};
 use homeboy_extension_contract::manifest_action_config::ActionConfig;
