+++ crates/homeboy-core/src/extension/lifecycle/install_sources.rs
@@ -15,7 +15,7 @@ use homeboy_core::git;
 use homeboy_core::paths;
 use homeboy_engine_primitives::local_files;
 
-use super::super::execution::run_setup;
+use super::super::invoke::run_setup;
 use super::{
     derive_id_from_url, manifest_path_for_extension, slugify_id, write_source_metadata,
     InstallResult,
