+++ crates/homeboy-core/src/extension/resolve.rs
@@ -5,7 +5,7 @@ use crate::error::{Error, ErrorCode, Result};
 use std::path::{Path, PathBuf};
 
 use crate::extension::catalog::{extension_path, load_extension, load_extension_in_root};
-use crate::extension_invocation_context::ResolvedExtensionInvocationContext;
+use crate::extension::invoke::ResolvedExtensionInvocationContext;
 use homeboy_extension_contract::ExtensionCapability;
 
 mod surface;
