+++ crates/homeboy-release/src/release/executor/publish.rs
@@ -52,7 +52,8 @@ pub(crate) fn run_publish(
         None,
         extra_config.as_ref(),
     );
-    let response = extension::execute_action(&extension.id, action_id, None, None, Some(&payload))?;
+    let response =
+        extension::invoke::execute_action(&extension.id, action_id, None, None, Some(&payload))?;
     let extension_data = serde_json::to_value(&response).map_err(|e| {
         Error::internal_json(e.to_string(), Some("extension action output".to_string()))
     })?;
